agrega y actualiza Libro de maneja exitosa

+++ src/main/java/edu/sucho/libreriaweb/service/AutorService.java
@@ -2,5 +2,5 @@ package edu.sucho.libreriaweb.service;
 
 import edu.sucho.libreriaweb.model.Autor;
 
-public interface AutorService extends BaseService<Autor, Long>{
+public interface AutorService extends BaseService<Autor, Integer>{
 }
