+++ src/main/java/edu/sucho/libreriaweb/service/EditorialService.java
@@ -2,5 +2,5 @@ package edu.sucho.libreriaweb.service;
 
 import edu.sucho.libreriaweb.model.Editorial;
 
-public interface EditorialService extends BaseService<Editorial, Long>{
+public interface EditorialService extends BaseService<Editorial, Integer>{
 }
