+++ src/main/java/edu/sucho/libreriaweb/service/LibroService.java
@@ -2,5 +2,5 @@ package edu.sucho.libreriaweb.service;
 
 import edu.sucho.libreriaweb.model.Libro;
 
-public interface LibroService extends BaseService<Libro, Long>{
+public interface LibroService extends BaseService<Libro, Integer>{
 }
